# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: SuriFleet API
  version: 0.1.0
  description: |
    REST API системы централизованного управления флотом Suricata IDS/IPS.
    Модель флота: организация → кластер → хост → инстанс Suricata (1..N на хост).
    Даты — RFC3339 (date-time). Идентификаторы — UUID.
    Идемпотентность мутирующих POST (деплои, сборка ruleset) — заголовок Idempotency-Key.
  contact:
    name: SuriFleet
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Единый префикс API (health/version — тоже здесь, без аутентификации)
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Локальный вход (email + пароль, в т.ч. break-glass)
      description: Только для локальных пользователей; SSO-пользователи входят через /auth/sso. Входы break-glass аудируются отдельно.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthTokens'}
        default: {$ref: '#/components/responses/Error'}
  /auth/refresh:
    post:
      operationId: refresh_tokens
      tags: [auth]
      summary: Обновить пару токенов по refresh-токену
      description: Refresh-токен ротируется; для SSO-пользователей здесь же проверяется активность в IdP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefreshRequest'}
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthTokens'}
        default: {$ref: '#/components/responses/Error'}
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Завершить сессию (отозвать refresh-токен)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefreshRequest'}
      responses:
        '204': {description: Сессия завершена}
        default: {$ref: '#/components/responses/Error'}
...
  /auth/me:
    get:
      operationId: get_current_user
      tags: [auth]
      summary: Текущий пользователь, его роли и итоговые разрешения
      responses:
        '200':
          description: Профиль и разрешения
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentUser'}
        default: {$ref: '#/components/responses/Error'}

  # --------------------------------------------------------------------------
  # Флот: организации, кластеры, хосты, инстансы
  # --------------------------------------------------------------------------
  /organizations:
    get:
      operationId: list_organizations
      tags: [organizations]
      summary: Список организаций
      description: "Требуется право: organizations.read"
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OrganizationPage'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: create_organization
      tags: [organizations]
      summary: Создать организацию
      description: "Требуется право: organizations.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrganizationInput'}
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
        default: {$ref: '#/components/responses/Error'}
  /organizations/{id}:
    parameters:
      - {$ref: '#/components/parameters/PathId'}
    get:
      operationId: get_organization
      tags: [organizations]
      summary: Карточка организации
      description: "Требуется право: organizations.read"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: update_organization
      tags: [organizations]
      summary: Изменить организацию
      description: "Требуется право: organizations.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrganizationUpdateInput'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Organization'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: delete_organization
      tags: [organizations]
      summary: Удалить организацию (каскадно всё содержимое)
      description: "Требуется право: organizations.write"
      responses:
        '204': {description: Удалено}
        default: {$ref: '#/components/responses/Error'}
  /clusters:
    get:
      operationId: list_clusters
      tags: [clusters]
      summary: Список кластеров
      description: "Требуется право: clusters.read"
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {name: organization_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ClusterPage'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: create_cluster
      tags: [clusters]
      summary: Создать кластер
      description: "Требуется право: clusters.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ClusterInput'}
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Cluster'}
        default: {$ref: '#/components/responses/Error'}
  /clusters/{id}:
    parameters:
      - {$ref: '#/components/parameters/PathId'}
    get:
      operationId: get_cluster
      tags: [clusters]
      summary: Карточка кластера
      description: "Требуется право: clusters.read"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Cluster'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: update_cluster
      tags: [clusters]
      summary: Изменить кластер
      description: "Требуется право: clusters.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ClusterUpdateInput'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Cluster'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: delete_cluster
      tags: [clusters]
      summary: Удалить кластер
      description: "Требуется право: clusters.write"
      responses:
        '204': {description: Удалено}
        default: {$ref: '#/components/responses/Error'}
...
  /hosts:
    get:
      operationId: list_hosts
      tags: [hosts]
      summary: Список хостов
      description: "Требуется право: hosts.read"
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {name: cluster_id, in: query, schema: {type: string, format: uuid}}
        - {name: q, in: query, description: Поиск по hostname/IP, schema: {type: string}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HostPage'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: create_host
      tags: [hosts]
      summary: Зарегистрировать хост вручную (обычно хост создаётся enrollment'ом)
      description: "Требуется право: hosts.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HostInput'}
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Host'}
        default: {$ref: '#/components/responses/Error'}
  /hosts/{id}:
    parameters:
      - {$ref: '#/components/parameters/PathId'}
    get:
      operationId: get_host
      tags: [hosts]
      summary: Карточка хоста (с агентом и инстансами)
      description: "Требуется право: hosts.read"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HostDetail'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: update_host
      tags: [hosts]
      summary: Изменить хост (hostname, labels)
      description: "Требуется право: hosts.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HostUpdateInput'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Host'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: delete_host
      tags: [hosts]
      summary: Удалить хост (агент отвязывается, инстансы удаляются)
      description: "Требуется право: hosts.write"
      responses:
        '204': {description: Удалено}
        default: {$ref: '#/components/responses/Error'}
...
  /instances:
    get:
      operationId: list_instances
      tags: [instances]
      summary: Список инстансов
      description: "Требуется право: hosts.read"
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {name: host_id, in: query, schema: {type: string, format: uuid}}
        - {name: cluster_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstancePage'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: create_instance
      tags: [instances]
      summary: Добавить инстанс вручную
      description: "Требуется право: hosts.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/InstanceInput'}
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Instance'}
        default: {$ref: '#/components/responses/Error'}
  /instances/{id}:
    parameters:
      - {$ref: '#/components/parameters/PathId'}
    get:
      operationId: get_instance
      tags: [instances]
      summary: Карточка инстанса
      description: "Требуется право: hosts.read"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Instance'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: update_instance
      tags: [instances]
      summary: Изменить инстанс (пути, интерфейсы, systemd unit)
      description: "Требуется право: hosts.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/InstanceUpdateInput'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Instance'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: delete_instance
      tags: [instances]
      summary: Удалить инстанс из-под управления
      description: "Требуется право: hosts.write"
      responses:
        '204': {description: Удалено}
        default: {$ref: '#/components/responses/Error'}

  # --------------------------------------------------------------------------
  # Правила
  # --------------------------------------------------------------------------
  /rules:
    get:
      operationId: list_rules
      tags: [rules]
      summary: Список правил мастер-репозитория
      description: "Требуется право: rules.read. Фильтры комбинируются (AND)."
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {name: status, in: query, schema: {type: string, enum: [enabled, disabled, expired, under_review, deleted]}}
        - {name: category, in: query, schema: {type: string}}
        - {name: tag, in: query, schema: {type: string}}
        - {name: source, in: query, description: 'Источник: file | feed | ioc | et_open | et_pro', schema: {type: string, enum: [file, feed, ioc, et_open, et_pro]}}
        - {name: sid, in: query, schema: {type: integer, format: int64}}
        - {name: q, in: query, description: Поиск по msg, schema: {type: string}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RulePage'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: create_rule
      tags: [rules]
      summary: Создать правило вручную
      description: "Требуется право: rules.write. Сырой текст парсится и валидируется; создаётся первая ревизия."
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RuleInput'}
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Rule'}
        default: {$ref: '#/components/responses/Error'}
  /rules/{id}:
    parameters:
      - {$ref: '#/components/parameters/PathId'}
    get:
      operationId: get_rule
      tags: [rules]
      summary: Карточка правила
      description: "Требуется право: rules.read"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Rule'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: update_rule
      tags: [rules]
      summary: "Изменить правило (тюнинг аналитика: status, priority, threshold, теги)"
      description: "Требуется право: rules.write. Тюнинг сохраняется между ревизиями фида."
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RuleUpdateInput'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Rule'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: delete_rule
      tags: [rules]
      summary: Удалить правило (перевод в статус deleted)
      description: "Требуется право: rules.write"
      responses:
        '204': {description: Удалено}
        default: {$ref: '#/components/responses/Error'}
  /rules/import:
    post:
      operationId: import_rules
      tags: [rules]
      summary: Импорт правил из файла или текста
      description: "Требуется право: rules.write. Парсинг и валидация при импорте; существующие sid обновляются новой ревизией, тюнинг аналитика сохраняется."
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: {type: string, format: binary, description: Файл .rules}
                category: {type: string, description: 'Override категории для всех правил файла (иначе — classtype)'}
                source: {type: string, enum: [file, feed], default: file}
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: {type: string, description: Содержимое .rules}
                category: {type: string, description: 'Override категории для всех правил (иначе — classtype)'}
                source: {type: string, enum: [file, feed], default: file}
      responses:
        '200':
          description: Статистика импорта
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ImportResult'}
        default: {$ref: '#/components/responses/Error'}
  /rules/bulk:
    post:
      operationId: bulk_rules
      tags: [rules]
      summary: Массовая операция над правилами
      description: "Требуется право: rules.write. Цель — список ids или фильтр; действия: enable | disable | delete | set_priority | add_tag | remove_tag."
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BulkRequest'}
      responses:
        '200':
          description: Число затронутых правил
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BulkResult'}
        default: {$ref: '#/components/responses/Error'}
...
  /feeds:
    get:
      operationId: list_feeds
      tags: [feeds]
      summary: Список фидов
      description: "Требуется право: feeds.read"
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {name: type, in: query, schema: {type: string, enum: [et_open, et_pro, taxii, stix, misp, generic]}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FeedPage'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: create_feed
      tags: [feeds]
      summary: Подключить фид
      description: "Требуется право: feeds.write. Креды (credentials) — write-only, хранятся шифрованно."
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FeedInput'}
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Feed'}
        default: {$ref: '#/components/responses/Error'}
  /feeds/{id}:
    parameters:
      - {$ref: '#/components/parameters/PathId'}
    get:
      operationId: get_feed
      tags: [feeds]
      summary: Карточка фида
      description: "Требуется право: feeds.read"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Feed'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: update_feed
      tags: [feeds]
      summary: Изменить фид
      description: "Требуется право: feeds.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FeedUpdateInput'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Feed'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: delete_feed
      tags: [feeds]
      summary: Отключить и удалить фид
      description: "Требуется право: feeds.write. Импортированные правила и IOC остаются (feed_id → NULL по FK), история запусков удаляется каскадом."
      responses:
        '204': {description: Удалено}
        default: {$ref: '#/components/responses/Error'}
...
  /iocs:
    get:
      operationId: list_iocs
      tags: [iocs]
      summary: Список IOC
      description: "Требуется право: iocs.read"
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {name: type, in: query, schema: {type: string, enum: [ip, domain, url, md5, sha1, sha256, email]}}
        - {name: status, in: query, schema: {type: string, enum: [active, under_review, expired, revoked]}}
        - {name: source, in: query, description: Источник (имя фида или ручной), schema: {type: string}}
        - {name: q, in: query, description: Поиск по значению, schema: {type: string}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/IocPage'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: create_ioc
      tags: [iocs]
      summary: Создать IOC вручную
      description: "Требуется право: iocs.write"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IocInput'}
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Ioc'}
        default: {$ref: '#/components/responses/Error'}
  /iocs/{id}:
    parameters:
      - {$ref: '#/components/parameters/PathId'}
    get:
      operationId: get_ioc
      tags: [iocs]
      summary: Карточка IOC
      description: "Требуется право: iocs.read"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Ioc'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      operationId: update_ioc
      tags: [iocs]
      summary: Изменить IOC (скоринг, статус, срок жизни)
      description: "Требуется право: iocs.write. Смена status на revoked отзывает сгенерированное из IOC правило (source_type=ioc → status=disabled + тег ioc-revoked)."
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IocUpdateInput'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Ioc'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: delete_ioc
      tags: [iocs]
      summary: Удалить IOC
      description: "Требуется право: iocs.write. Жёсткое удаление; сгенерированное из IOC правило отзывается (status=disabled + тег ioc-revoked)."
      responses:
        '204': {description: Удалено}
        default: {$ref: '#/components/responses/Error'}
...
  /iocs/import:
    post:
      operationId: import_iocs
      tags: [iocs]
      summary: Импорт IOC списком (JSON)
      description: "Требуется право: iocs.write. Идемпотентно по (org, type, value): новые — imported, существующие — updated (score/source/expires_at; status не перетирается). Ошибочные элементы не прерывают импорт."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                source: {type: string, description: Источник по умолчанию для элементов без своего source}
                items:
                  type: array
                  items: {$ref: '#/components/schemas/IocInput'}
      responses:
        '200':
          description: Статистика импорта
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ImportResult'}
        default: {$ref: '#/components/responses/Error'}

  # --------------------------------------------------------------------------
  # Деплой и состояние (ключевое требование А)
  # --------------------------------------------------------------------------
...
  /rulesets/{id}:
    parameters:
      - {$ref: '#/components/parameters/PathId'}
    get:
      operationId: get_ruleset
      tags: [rulesets]
      summary: Карточка версии ruleset (состав, манифест)
      description: "Требуется право: deployments.read"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RulesetVersionDetail'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Сессионный токен (непрозрачный; MVP чанка 28 — access и refresh совпадают, refresh ротирует; JWT — позже)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API-токен автоматизации (scopes при выпуске)
  parameters:
    Limit:
      name: limit
      in: query
      description: Размер страницы (keyset)
      schema: {type: integer, default: 100, minimum: 1, maximum: 1000}
...
  responses:
    Error:
      description: Единый формат ошибки
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: {type: string}
    AuthTokens:
      type: object
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        token_type: {type: string, example: Bearer}
        expires_in: {type: integer, description: "Срок жизни access token, секунды"}
...